request handlers removed
