+++ sdk/tables/azure-data-tables/azure/data/tables/_base_client.py
@@ -41,7 +41,7 @@ from azure.core.pipeline.policies import (
 )
 
 from .shared_access_signature import QueryStringConstants
-from ._shared._constants import STORAGE_OAUTH_SCOPE, SERVICE_HOST_BASE, CONNECTION_TIMEOUT, READ_TIMEOUT
+from ._constants import STORAGE_OAUTH_SCOPE, SERVICE_HOST_BASE, CONNECTION_TIMEOUT, READ_TIMEOUT
 from ._shared.models import LocationMode
 from ._authentication import SharedKeyCredentialPolicy
 from ._policies import (
